implemented Pyomo solver for solving
implemented a new function script solver_config.py that handes creating a pyomo solver for the specified solver specified in the solver_settings.json

changed from gurobi to pyomo:
- opti_central
- opti_dimensioning_cenrtral_devices
- k_medoids
- optimizer

implemented for debugging purposes that the model is writen in the results/ folder. If an error occures an errorfile is created.